+++ src/main/java/de/samply/db/repository/ProjectCoderRepository.java
@@ -1,6 +1,6 @@
 package de.samply.db.repository;
 
-import de.samply.db.model.Project;
+import de.samply.db.model.ProjectBridgeheadUser;
 import de.samply.db.model.ProjectCoder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,6 +10,8 @@ import java.util.Optional;
 @Repository
 public interface ProjectCoderRepository extends JpaRepository<ProjectCoder, Long> {
 
-    Optional<ProjectCoder> findByProjectAndEmail (Project project, String email);
+    Optional<ProjectCoder> findByProjectBridgeheadUser(ProjectBridgeheadUser projectBridgeheadUser);
+
+    Optional<ProjectCoder> findByProjectBridgeheadUserAndDeletedAtIsNull(ProjectBridgeheadUser projectBridgeheadUser);
 
 }