+++ src/main/java/de/samply/project/ProjectType.java
@@ -2,7 +2,6 @@ package de.samply.project;
 
 public enum ProjectType {
     EXPORT,
-    DATASHIELD
-    //,
-    //RESEARCH_ENVIRONMENT
+    DATASHIELD,
+    RESEARCH_ENVIRONMENT
 }